requirements.txt
Click>=7.0
Pillow>=5.4.1
pendulum>=2.0.4
python-magic>=0.4.15